Authentication

+++ db/db_user.py
@@ -2,6 +2,7 @@ from sqlalchemy.orm.session import Session
 from db.hash import Hash
 from schemas import UserBase
 from db.models import DbUser
+from fastapi import HTTPException, status
 
 def create_user(db: Session, request: UserBase):
   new_user = DbUser(
@@ -20,11 +21,26 @@ def get_all_users(db: Session):
 
 def get_user(db: Session, id: int):
   #Handle any excemptions
-  return db.query(DbUser).filter(DbUser.id == id).first()
+  user = db.query(DbUser).filter(DbUser.id == id).first()
+  if not user:
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+    detail=f'User with id {id} not found')
+  return 
+
+def get_user_by_username(db: Session, username: str):
+  #Handle any excemptions
+  user = db.query(DbUser).filter(DbUser.username == username).first()
+  if not user:
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+      detail=f'User with username {username} not found')
+  return user
 
 def update_user(db: Session, id: int, request: UserBase):
   user = db.query(DbUser).filter(DbUser.id == id)
   #Handle any excemptions
+  if not user.first():
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+    detail=f'User with id {id} not found')
   user.update({
     DbUser.username: request.username,
     DbUser.email: request.email,
@@ -34,8 +50,11 @@ def update_user(db: Session, id: int, request: UserBase):
   return 'ok'
 
 def delete_user(db: Session, id: int):
-  user = db.query(DbUser).filter(DbUser.id == id).first()
+  user = db.query(DbUser).filter(DbUser.id == id)
   #Handle any excemptions
+  if not user:
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+    detail=f'User with id {id} not found')
   db.delete(user)
   db.commit()
   return 'ok'
